Fixes when transferring large buffers

+++ Adafruit_MQTT.h
@@ -292,7 +292,7 @@ private:
   // Functions to generate MQTT packets.
   uint8_t connectPacket(uint8_t *packet);
   uint8_t disconnectPacket(uint8_t *packet);
-  uint16_t publishPacket(uint8_t *packet, const char *topic, uint8_t *payload,
+  uint32_t publishPacket(uint8_t *packet, const char *topic, uint8_t *payload,
                          uint32_t bLen, uint8_t qos, uint32_t maxPacketLen = 0);
   uint8_t subscribePacket(uint8_t *packet, const char *topic, uint8_t qos);
   uint8_t unsubscribePacket(uint8_t *packet, const char *topic);
